requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
google==3.0.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.10
numpy==1.26.4
packaging==23.1
Pillow==9.5.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
pytesseract==0.3.10
requests==2.32.3
rsa==4.9
setuptools==75.1.0
soupsieve==2.6
uritemplate==4.1.1
urllib3==2.2.3
